refined the data flow, identified errors in the data flow, including db state overwriting local state when multiple tree instances are sending their specific info to the db and the app state is being refreshed from all the habits there before the other tree instances can update their stuff, and the new app state flows those props into the tree updating the tree incorrectly in the case of the dailyComplete.
The quick fix results in congruency in local state and db state but results in 2x calls to the db and renders.  There is another way.  This involves creating a new api to handle sending all habits to db.  Then from tree, we can call a ajax request to the db merging the tree habit with the rest of habits in state and sending them to the db.  This will reduce renders and db calls 1/2, however at the price of more bugs possibly, like making sure having to make daily complete false in the tree useEffect bc its not making it to the db from the checkbox.

Next task will be the shrinking and R3F tooltip.  Then we are at a working demo.